Adjusting frames and tests to now use the new FrameType type.

+++ internal/wire/reset_stream_frame.go
@@ -56,9 +56,9 @@ func parseResetStreamFrame(b []byte, isResetStreamAt bool, _ protocol.Version) (
 
 func (f *ResetStreamFrame) Append(b []byte, _ protocol.Version) ([]byte, error) {
 	if f.ReliableSize == 0 {
-		b = quicvarint.Append(b, ResetStreamFrameType)
+		b = quicvarint.Append(b, uint64(ResetStreamFrameType))
 	} else {
-		b = quicvarint.Append(b, ResetStreamAtFrameType)
+		b = quicvarint.Append(b, uint64(ResetStreamAtFrameType))
 	}
 	b = quicvarint.Append(b, uint64(f.StreamID))
 	b = quicvarint.Append(b, uint64(f.ErrorCode))
